requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.10
passlib==1.7.4
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
